+++ lib/core/utils/clock.dart
@@ -1,0 +1,11 @@
+import 'package:injectable/injectable.dart';
+
+abstract class Clock {
+  DateTime now();
+}
+
+@LazySingleton(as: Clock)
+class SystemClock implements Clock {
+  @override
+  DateTime now() => DateTime.now();
+}