+++ test/unit/domain/usecases/create_health_log_test.dart
@@ -1,6 +1,7 @@
 import 'package:dartz/dartz.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:health_tracker_reports/core/error/failures.dart';
+import 'package:health_tracker_reports/core/utils/clock.dart';
 import 'package:health_tracker_reports/domain/entities/health_log.dart';
 import 'package:health_tracker_reports/domain/entities/reference_range.dart';
 import 'package:health_tracker_reports/domain/entities/vital_measurement.dart';
@@ -17,6 +18,8 @@ class MockValidateVitalMeasurement extends Mock
     implements ValidateVitalMeasurement {}
 
 class MockUuid extends Mock implements Uuid {}
+
+class MockClock extends Mock implements Clock {}
 
 void main() {
   late MockHealthLogRepository mockRepository;
@@ -44,16 +47,19 @@ void main() {
     mockRepository = MockHealthLogRepository();
     mockValidate = MockValidateVitalMeasurement();
     mockUuid = MockUuid();
+    final mockClock = MockClock();
 
     var callIndex = 0;
     final generatedIds = ['log-id', 'vital-1', 'vital-2', 'vital-3'];
     when(() => mockUuid.v4()).thenAnswer((_) => generatedIds[callIndex++]);
 
+    when(() => mockClock.now()).thenReturn(fixedNow);
+
     usecase = CreateHealthLog(
       repository: mockRepository,
       validateVitalMeasurement: mockValidate,
       uuid: mockUuid,
-      now: () => fixedNow,
+      clock: mockClock,
     );
   });